+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/Eyes.java
@@ -161,7 +161,7 @@ public class Eyes extends EyesBase {
 
     @Override
     public String getBaseAgentId() {
-        return "eyes.selenium.java/4.13.3";
+        return "eyes.selenium.java/4.14.0";
     }
 
     public EyesWebDriver getEyesDriver() {